Schema: support multiple Schema nodes
The $metadata may contain one or more Schema nodes. Each Schema node
must have own Namespace.

I am not sure how to deal with conflicts in EntitySet names. The problem
is that EntitySet name is a part URL and it is used without Namespace.

For now, I ignore the problem described above.