Handle association set ends with same entity sets
In the metadata document, it is allowed to declare same Entity Sets in
different ends of one Association Set. Therefore, we cannot utilize
dictionary to store such ends (values can be overwritten).

In this commit, there is added a new class AssociationSetEndRole which
represents one end within the Association Set. End roles are stored in
a list, in the same manner like end roles in the Association. Also,
each end now contains a reference to a corresponding entity set.