Remove global variables in favor of configuration class
This commit removes global variables and all issues tied to them. We
decided to replace global vars with instance of class "Config" which
works as an envelope for variables which should be parser-wide available.

Class EDMX is deprecated in favor of MetadataBuilder which takes instance
of config as parameter. Param config is optional.