Fix datetime handling while creating/updating entity
For most of EDM types, their value and json representation are the same.
One of the types with different behaviour is DateTime which is in json
in OData2 represented as "/Data(xxxx)/" instead of simply timestamp
xxxx. Thus, the need for 'to_json' function which for most data types
solely returns value. However, for DateTime, it returns a string in a
format mentioned above.

Reference links:

For literal format see "6. Primitive Data Types" at
https://www.odata.org/documentation/odata-version-2-0/overview/

For json format definition see "4. Primitive Types" at
https://www.odata.org/documentation/odata-version-2-0/json-format/