client: add the possibility to provide static metadata
Sometimes we have to deal with bug reports where the reporter provides
us with their metadata file and I want to add the possibility to skip
downloading of metadata into my test tool at `bin/pyodata`.

We also discussed the possibility do implement this feature in #24.

It is true that pyodata users can create an instance of Service even
now:

```python
   schema = pyodata.v2.model.MetadataBuilder(metadata).build()
   service = pyodata.v2.service.Service(url, schema, connection)
```

But they would not get the same results because we initialize
MetadataBuilder with config.

So, I want to enhance Client in order to avoid future problems and
to make it convenient.

Usage:

Since we use lxml, we must provide bytes (it will not work with str).

```python
   with open('/the/file/path.xml', 'rb') as thefile:
       metadata = thefile.read()
```

Notice 'b'.