model: implement floating point numbers handling
It turned out that PyOData does not understand floating point numbers as
defined in OData standards.

This commit adds special formatters and parsers for floating point
numbers as represented in OData standards.

The commit also adds the type Edm.Float which is not defined in
the OData V2 standard but it is most probably of a new created
Edm format specification.

The OData standard is kinda fuzzy and even examples do not match
the requirements. Hence, do not be confused by the fact that we do not
add the suffixes {d,f} to string representations and we don't check if
converted literals have the suffix.

Closes #81