Replace python builtin exceptions with custom ones
This commit is divided into these subtasks:
1. Categorizing all exceptions into groups[1][2]
2. Creating custom exceptions(all are annotated directly in file exceptions.py)
3. Replacing all raise EXCEPTION with newly created ones
4. Resolving dependent try-catch block, i. e. replace builtin exceptions in catch statements.

The 4th task is a bit tricky. I had to change the code a bit more to reflect things such as certain catch block being redundant(model line 1170, 1074-1078, 1155), multiple previously different exceptions being now caught as one(model lines 1191-1210)

Tests have been changed accordingly. In a few places, assert_logging_policy was replaced by caplog. It is a preferable practice and doing so made it easier to adjust the test.

Also, a bug causing the test silently passing when it should not have was found on the line 1245-1250.

[1] https://github.com/mamiksik/python-pyodata/wiki/Exceptions
[2] https://github.com/mamiksik/python-pyodata/wiki/Exceptions-rewrite