tree-wide: correctly handle null values
Only a declared property knows if it accepts Null or not. I was playing
with nullable Type but than I would have to have also nullable
TypeTraits and there result would be a complete mess. Therefore I decided
to move the responsibility to serialize/deserialize the declared
property. The decision led to a simpler code which is a good sign.

I could have used a better exception type but we currently evaluating
the best exception strategies, so I used a generic type.

Part of the issue #95