bin: enable reading of EntitySets per partes
Some services limit size of GET EntitySet response to a certain number
of Entities (20, 1000, ...). This patch adds the possibility to specify
number of entities retrieved in a request.

The executable utility will try to get all entities in one request
and will not check number of returned entities by default.

If the user adds '--chunk-size 5', the utility will be sending GET
requests until it read the entity set completely.