Fix incorrect length when creating an HTTPResponse from FakeSocket
Due to the response_str encountering inadvertent mutation elsewhere in
the code, the size slightly changes and with a large response, does not
correctly get reported to HTTPResponse.

We manually specify the length of the response to the size of the
BytesIO buffer and this gives HTTPResponse enough room to read to the
end of the response.

The ODataHttpResponse.from_string method fails because in the process of
converting the multipart response to and from bytes and also being
passed around, some bytes are being added somewhere and causing the size
of the data to no longer be equal to Content-Type.

This causes HTTPResponse.begin() to seed the wrong length attribute,
which causes the file pointer to not read far enough which, in
turn, causes a JSON parse error. So to clean this up we just need
to manually set response.length to how large the body is now after
all its encoding and decoding.

Signed-off-by: Jakub Filak <jakub.filak@sap.com>