service: add etag property to EntityProxy
In OData V2 ETag is located in header or is part of the body in optional
substructure called "__metadata".

https://www.odata.org/documentation/odata-version-2-0/json-format/#7_representing_entries_8

ETag is included in both the header and body of response we check if
they match. If not exception is raised which is our invention and
does not need to be preserved if it causes any problems.