disable pylint logging-format-interpretation for f-strings
This one is basically a warning for potentional performance loss,
for scenarios when log message is not really logged. With 3 instances
of such problem in pyodata and the perf loss being neglegable, IMHO is better
to not mix old and new style of formatting. Therefore disabling the warning.

See more:
https://github.com/PyCQA/pylint/issues/2395
https://github.com/PyCQA/pylint/issues/1788
https://stackoverflow.com/questions/34619790/pylint-message-logging-format-interpolation